#1: adding edit into main flows command
